// packages/cli/src/ui/types.ts
export enum MessageType {
  INFO = 'info',
  ERROR = 'error',
  USER = 'user',
  ABOUT = 'about',
  HELP = 'help',
  STATS = 'stats',
  MODEL_STATS = 'model_stats',
  TOOL_STATS = 'tool_stats',
  QUIT = 'quit',
  QUIT_CONFIRMATION = 'quit_confirmation',
  GEMINI = 'gemini',
  COMPRESSION = 'compression',
  SUMMARY = 'summary',
}